Modify SYSTEMIO APIs parameter and fix build error

+++ framework/include/iotbus/iotbus_spi.h
@@ -93,7 +93,7 @@ iotbus_spi_context_h iotbus_spi_open(unsigned int bus, const struct iotbus_spi_c
  * @return On success, 0 is returned. On failure, a negative value is returned.
  * @since Tizen RT v1.0
  */
-int iotbus_spi_write(iotbus_spi_context_h hnd, uint8_t *txbuf, int length);
+int iotbus_spi_write(iotbus_spi_context_h hnd, uint8_t *txbuf, size_t length);
 
 /**
  * @brief reads data over spi bus.
@@ -104,7 +104,7 @@ int iotbus_spi_write(iotbus_spi_context_h hnd, uint8_t *txbuf, int length);
  * @return On success, 0 is returned. On failure, a negative value is returned.
  * @since Tizen RT v1.0
  */
-int iotbus_spi_recv(iotbus_spi_context_h hnd, uint8_t *rxbuf, int length);
+int iotbus_spi_recv(iotbus_spi_context_h hnd, uint8_t *rxbuf, size_t length);
 
 /**
  * @brief transfers rx and tx data over spi bus.
@@ -116,7 +116,7 @@ int iotbus_spi_recv(iotbus_spi_context_h hnd, uint8_t *rxbuf, int length);
  * @return On success, 0 is returned. On failure, a negative value is returned.
  * @since Tizen RT v1.0
  */
-int iotbus_spi_transfer_buf(iotbus_spi_context_h hnd, uint8_t *txbuf, uint8_t *rxbuf, int length);
+int iotbus_spi_transfer_buf(iotbus_spi_context_h hnd, uint8_t *txbuf, uint8_t *rxbuf, size_t length);
 
 /**
  * @brief closes spi_context.
